pyproject.toml
[project]
name = "ai-news-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.5.0.post8",
    "fastapi>=0.115.12",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
